Updated the plot.heat function
The signature for the plot.heat function now contains a cluster merge
boolean and a field for the merging radius. The function also has been
modified to save an unaltered copy of the unified distance matrix and if
the merge switch is thrown pass it along to the cluster merging
functions.